AAAAAH

+++ TF - Administrador de recibos por honorarios/administrador_recibos/src/main/java/com/finanzas/administrador_recibos/controller/FactoringController.java
@@ -1,5 +1,6 @@
 package com.finanzas.administrador_recibos.controller;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,9 +10,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.finanzas.administrador_recibos.model.DetalleFactoring;
 import com.finanzas.administrador_recibos.model.Factoring;
-import com.finanzas.administrador_recibos.model.ReciboHonorarios;
 import com.finanzas.administrador_recibos.service.FactoringService;
 
 @RestController
@@ -34,7 +33,7 @@ public class FactoringController {
 	public Factoring registrarFactoring(@RequestBody Factoring factoring,
 			@RequestParam(name = "tipoTasa", required = true) String tipoTasa, 
 			@RequestParam(name = "capitalizacion", required = false) String capitalizacion, 
-			@RequestParam(name = "pSegDesg", required = true) double pSegDesg, 
+			@RequestParam(name = "pSegDesg", required = true) BigDecimal pSegDesg, 
 			@RequestParam(name = "tipoMoneda", required = true) String tipoMoneda) throws Exception{
 		
 		return this.factoringService.registrarFactoring(factoring, tipoTasa, capitalizacion, pSegDesg, tipoMoneda);
